Refactor `EscapeUrl` function and usage
- Use `size_t` for size argument.
- Use templated function when possible.
- Use `str_startswith` instead of hard-coding offset.

+++ src/engine/shared/http.cpp
@@ -51,7 +51,7 @@ static int CurlDebug(CURL *pHandle, curl_infotype Type, char *pData, size_t Data
 	return 0;
 }
 
-void EscapeUrl(char *pBuf, int Size, const char *pStr)
+void EscapeUrl(char *pBuf, size_t Size, const char *pStr)
 {
 	char *pEsc = curl_easy_escape(nullptr, pStr, 0);
 	str_copy(pBuf, pEsc, Size);
